### 2019-05-06 #1433 @whedon
[PRE REVIEW]: CSaransh: Software Suite to Study Molecular Dynamics Simulations of Collision Cascades
**Submitting author:** @haptork (<a href="http://orcid.org/0000-0002-1100-6059">Utkarsh Bhardwaj</a>)
**Repository:** <a href="https://github.com/haptork/csaransh" target ="_blank">https://github.com/haptork/csaransh</a>
**Version:** v0.3.1
