### 2019-05-09 #1433 @katyhuff
@arose Alexander Rose, are you able to and interested in taking on this review?
This submission is in need of expertise in physics, in particular, molecular dynamics related to radiation damage cascades, as well as familiarity with the systems and languages involved: python and c++ as well as, ideally, node.js. I know that you have at least a subset of these areas of expertise. I hope you'll let me know if you're able and willing to review it.

**Title:** CSaransh: Software Suite to Study Molecular Dynamics Simulations of Collision Cascades
**Summary:** The software suite to post-process, explore and visualize Molecular Dynamics (MD) simulations of collision cascades. It is an elaborate software solution for studing MD results of radiation damage simulations, starting from identifying defects from xyz file to finding correlations, visualizing subcascades to pattern matching clusters and the list goes on. There are many novel methods. The algorithms employed are fast and there is a refreshing interactive webapp to explore the results. You can check the web page for the project (link) and read the manual to go through the different sections of the analysis. The web page shows the results for the MD database of IAEA Challenge on Materials for Fusion 2018.
**Article Proof:** https://github.com/openjournals/joss-papers/blob/joss.01433/joss.01433/10.21105.joss.01433.pdf
**Submitting author:** @haptork (Utkarsh Bhardwaj)
**Repository:** https://github.com/haptork/csaransh
**Version:** v0.3.1
