### 2019-05-09 #1433 @jmborr
Hi Katy,

I have looked onto the repository, this project is 90% javascript, 8% C++
and 2% python. I am proficient with python, average with C++ and a total
ignorant with javascript. Thus, I am pretty sure I can't make a proper
review of this package. 😞

<jose/>
