### 2019-05-09 #1433 @katyhuff
@jmborr I had the same reaction initially, actually -- It seems to me, though, that all the science is in the post processor (which is solely in this directory: https://github.com/haptork/csaransh/tree/master/csaransh-pp) . The javascript seems to be separate from the science, and is more like a support utility. If you don't mind following the installation instructions in the readme, then perhaps you could just reveiw the scientific contributions in the postprocessor? The rest, I think just supports the interface. @haptork : can you confirm that this is the right interpretation?
