### 2019-05-10 #1433 @haptork
@jmborr @katyhuff yes, the algorithms and methods are in the post-processor directory mentioned by katyhuff. The post-processor writes the results it finds in JSON format which are then visualised and presented by web interface, written using javascript present in its entirety in csaransh-server directory.

The C++ code includes algorithms for finding the defects from xyz file of the MD simulation, finding features to characterise the clusters, find similar clusters etc. The python code adds machine learning results such as dimensionality of clusters and cascades using PCA, number of sub-cascades using DBSCAN, dimensionality reduction on features of clusters and their classification using UMAP, t-SNE and HDBSCAN etc.
