### 2019-05-10 #1433 @jmborr
I could take a look to the c++/python source, but  then I was wondering
what best software practices are required for all submissions. Right off
the bat I don't see any tests (unit, integration, or otherwise). From this
fact I assume there's no kind of continuous integration either.
This would be my first review so  I ignore if these practices are a must.
<jose/>
