### 2019-05-13 #1433 @katyhuff
Thanks both (@jmborr and @arose).

Unfortunately - @jmborr has made an excellent point that I hadn't noticed at first glance. @haptork - I do not see any testing here. There appear to be no unit tests and while the repository holds some example input files (and, perhaps, corresponding output) these do not seem to constitute integration tests. Can you comment on this? Tests are a requirement among the many features we expect in a JOSS submission. From the [author guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html):

> Tests
> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works

If you would like to withdraw your submission for the time being to work on these requirements, please let me know.
