### 2019-05-15 #1433 @ManojWarrier
As co-author, these are my comments:
1) The physics and results from CSaransh are described in the following paper -> https://arxiv.org/abs/1811.10923
2) Regarding tests,
    (a) CSaransh reproduces published physics results for the "number of defects" as a function of "primary knock-on atom energy".
    (b) The part where a certain type of cluster is chosen and other similar clusters are identified using pattern matching is a test in itself. One sees that similar clusters are identified.

Utkarsh, please decide on including a test suite where the above, and anything else besides, you find appropriate is included.
