### 2019-05-15 #1433 @haptork
@katyhuff The software includes sample inputs in the data directory as you noted. The installation steps in the manual include the steps to run the post-processors on the sample data given and view the results using the interface. I think this passes the "OK" guideline for the tests.

Different results including number of defects, cluster shapes have been verified qualitatively against prior published results and results found using other algorithms while many new results can also be verified qualitatively since the software includes visualisation tools as mentioned by @ManojWarrier .

We understand the importance of unit-tests and continuous integration especially as the software development progresses. We will work on making a test-suite for unit-testing possibly after a few more interesting research explorations. However, I believe that the software passes the "OK" guideline as is, since it includes the manual steps to run it on the included sample data and checking the behaviour by loading the processed data for visualisation. In addition to the sample data, the link to IAEA challenge database is included if someone wants to do more elaborate behavioural testing on 76 cascades at different energies and different elements.
