### 2019-05-15 #1433 @katyhuff
Thank you @haptork and @ManojWarrier .

From what I can tell, when you say "The installation steps in the manual include the steps to run the post-processors on the sample data given and view the results using the interface. "  you are referring to this statement:

> Run command should look something like this: %PATH%/csaransh_pp data/*/*xyz (this works on the sample data given with the repository).

While that's great, it doesn't constitute instructions for making an objective assertion of the software capabilities.

Our requirement for "ok" with regard to testing is that there must be an **objective assertion** of the correct behavior. When we say assertion, we do mean the "assert" function, or some manual instructions to make an objective comparison of observed results to expected results. Example input files are not tests if the user must determine the accuracy of the results qualitatively and with no guidance about what to expect as a correct result from the instructions.

To be clear, what I'm saying is that when one runs the example files, they'll see results (ideally) but will have to determine for themselves whether the results are correct. On their own, without an assertion command or a correct, expected result to compare against, these are not tests. This is clear from your description: you've used the word "qualitative" while our requirement for "ok" is this objective assertion. The following does not qualify as objective assertion.

> checking the behaviour by loading the processed data for visualisation.

Nor does the unguided, qualitative process you have described:

> Different results including number of defects, cluster shapes have been verified qualitatively against prior published results and results found using other algorithms while many new results can also be verified qualitatively since the software includes visualisation tools as mentioned by @ManojWarrier .

As the editor of this submission, I can start the review (with the help of our volunteer reviewers above). But, the submission will be stalled until there are tests that meet the criteria. At the moment, I would classify the current capabilities as "Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works." My recommendation is that you instead postpone your submission until the submission is more likely to meet our review requirements.
