### 2019-05-18 948b3e7
Added tests and refactored
