+++ csaransh-pp/test/defectsTests.cpp
@@ -1,0 +1,1056 @@
+#include <catch.hpp>
+
+#include <AddOffset.hpp>
+#include <xyz2defects.hpp>
+#include <results.hpp>
+
+using namespace csaransh;
+
+SCENARIO("Find nearest lattice site for a coordinate given lattice structure - "
+         "Addoffset",
+         "[defectsTest]") {
+  SECTION("Normal Cases - With 0.0 as origin") {
+    auto origin = Coords{{0.0, 0.0, 0.0}};
+    // Case 1
+    AddOffset a{1.0, "bcc", origin};
+    auto res = a(origin);
+    REQUIRE(std::get<0>(res)[0] == Approx(0.0));
+    REQUIRE(std::get<0>(res)[1] == Approx(0.0));
+    REQUIRE(std::get<0>(res)[2] == Approx(0.0));
+    REQUIRE(std::get<1>(res) == Approx(0.0));
+    REQUIRE(std::get<2>(res) == origin);
+    // Case a
+    auto resa = a(Coords{{0.4, 0.5, 0.5}});
+    REQUIRE(std::get<0>(resa)[0] == Approx(0.5));
+    REQUIRE(std::get<0>(resa)[1] == Approx(0.5));
+    REQUIRE(std::get<0>(resa)[2] == Approx(0.5));
+    REQUIRE(std::get<1>(resa) == Approx(0.1));
+    // Case b
+    auto cb = Coords{{2.0, 1.5, 3.71}};
+    auto resb = a(cb);
+    CHECK(std::get<0>(resb)[0] == Approx(2.5));
+    CHECK(std::get<0>(resb)[1] == Approx(1.5));
+    CHECK(std::get<0>(resb)[2] == Approx(3.5));
+    // checking if the lattice site found is nearer than other sites around
+    CHECK(std::get<1>(resb) <= calcDist(cb, Coords{{2.0, 1.0, 3.0}}));
+    CHECK(std::get<1>(resb) <= calcDist(cb, Coords{{2.0, 2.0, 3.0}}));
+    CHECK(std::get<1>(resb) <= calcDist(cb, Coords{{2.0, 2.0, 4.0}}));
+    CHECK(std::get<1>(resb) <= calcDist(cb, Coords{{2.0, 1.0, 4.0}}));
+    CHECK(std::get<1>(resb) <= calcDist(cb, Coords{{1.5, 1.5, 3.5}}));
+    // Case c
+    auto cc = Coords{{-0.1, -0.3, -0.71}};
+    auto resc = a(cc);
+    CHECK(std::get<0>(resc)[0] == Approx(0.0));
+    CHECK(std::get<0>(resc)[1] == Approx(0.0));
+    CHECK(std::get<0>(resc)[2] == Approx(-1.0));
+    CHECK(std::get<1>(resc) <= calcDist(cc, Coords{{0.0, 0.0, 0.0}}));
+    CHECK(std::get<1>(resc) <= calcDist(cc, Coords{{-0.5, -0.5, -0.5}}));
+    CHECK(std::get<1>(resc) <= calcDist(cc, Coords{{0.0, 0.0, -1.0}}));
+    CHECK(std::get<1>(resc) == Approx(0.429).epsilon(0.001));
+    // Case d
+    auto cd = Coords{{-0.3, -0.3, -0.71}};
+    auto resd = a(cd);
+    CHECK(std::get<0>(resd)[0] == Approx(-0.5));
+    CHECK(std::get<0>(resd)[1] == Approx(-0.5));
+    CHECK(std::get<0>(resd)[2] == Approx(-0.5));
[…]
