+++ csaransh-pp/src/printJson.cpp
@@ -2,10 +2,11 @@
 #include <cluster2features.hpp>
 #include <results.hpp>
 
-void printClusterIds(const std::unordered_map<int, std::vector<int>> clusters, std::ofstream& outfile) {
+void printClusterIds(const std::unordered_map<int, std::vector<int>> clusters,
+                     std::ofstream &outfile) {
   size_t i = 0;
   for (const auto &it : clusters) {
-    outfile << '"' << it.first << "\":"; 
+    outfile << '"' << it.first << "\":";
     outfile << "[";
     csaransh::writeVector(it.second, outfile);
     outfile << "]";
@@ -15,10 +16,11 @@ void printClusterIds(const std::unordered_map<int, std::vector<int>> clusters, s
   }
 }
 
-void printClusterIVs(const std::unordered_map<int, int> clusters, std::ofstream& outfile) {
+void printClusterIVs(const std::unordered_map<int, int> clusters,
+                     std::ofstream &outfile) {
   size_t i = 0;
   for (const auto &it : clusters) {
-    outfile << '"' << it.first << "\":"; 
+    outfile << '"' << it.first << "\":";
     outfile << it.second;
     if (i != clusters.size() - 1) outfile << ", ";
     outfile << "\n";
@@ -26,7 +28,7 @@ void printClusterIVs(const std::unordered_map<int, int> clusters, std::ofstream&
   }
 }
 
-void printSingleFeat(const csaransh::featT &feats, std::ofstream &outfile) {
+void printSingleFeat(const csaransh::featT &feats, std::ostream &outfile) {
   outfile << "\"dist\": ";
   outfile << "[";
   csaransh::writeStdAr(std::get<0>(feats), outfile);
@@ -41,10 +43,11 @@ void printSingleFeat(const csaransh::featT &feats, std::ofstream &outfile) {
   outfile << "]\n";
 }
 
-void printFeats(const std::unordered_map<int, csaransh::featT> &feats, std::ofstream &outfile) {
+void printFeats(const std::unordered_map<int, csaransh::featT> &feats,
+                std::ostream &outfile) {
   size_t count = 0;
   for (const auto &it : feats) {
-    outfile << '"' << it.first << "\":"; 
+    outfile << '"' << it.first << "\":";
     outfile << "{";
     printSingleFeat(it.second, outfile);
     outfile << "}";
@@ -54,49 +57,58 @@ void printFeats(const std::unordered_map<int, csaransh::featT> &feats, std::ofst
 }
 
 auto strSimulationCode(csaransh::SimulationCode code) {
-  return (code == csaransh::SimulationCode::parcas) ? "parcas" : 
-             ((code == csaransh::SimulationCode::lammps) ? "lammps-xyz" : "lammps-disp");
+  return (code == csaransh::SimulationCode::parcas)
+             ? "parcas"
+             : ((code == csaransh::SimulationCode::lammps) ? "lammps-xyz"
+                                                           : "lammps-disp");
 }
 
-void csaransh::printJson(std::ofstream &outfile, const std::string &s, const csaransh::Info &i, int id, const int &nDefects, const int &nClusters, 
-               const int &maxClusterSizeI, const int &maxClusterSizeV, const double &inClusterFractionI, 
-               const double &inClusterFractionV, const csaransh::DefectVecT &defects, 
-               const std::array<std::vector<double>, 2> &dists, const std::array<std::vector<double>, 2> &angles, 
[…]
