+++ csaransh-pp/include/cluster2features.hpp
@@ -14,15 +14,14 @@
 #include <helper.hpp>
 
 namespace csaransh {
-//using defectsT = std::vector<std::tuple<Coords, bool, int, bool>>;
+// using defectsT = std::vector<std::tuple<Coords, bool, int, bool>>;
 
 using distFeatT = std::array<double, 40>;
 using angleFeatT = std::array<double, 36>;
-//using angleFeatT = std::array<double, 180>;
+// using angleFeatT = std::array<double, 180>;
 using adjNn2FeatT = std::array<double, 20>;
-using featT = std::tuple<
-  distFeatT, angleFeatT, adjNn2FeatT
->;
-featT pairHists(const std::vector<std::array<double, 3>> &v, const std::vector<bool> &v2, double latConst);
+using featT = std::tuple<distFeatT, angleFeatT, adjNn2FeatT>;
+featT pairHists(const std::vector<std::array<double, 3>> &v,
+                const std::vector<bool> &v2, double latConst);
 }
-#endif //CLUSTER2FEATURES_CSARANSH_HPP
+#endif // CLUSTER2FEATURES_CSARANSH_HPP

[…]
