### 2019-05-18 c92ac46
added conda environment file
