### 2019-05-18 003095a
Added travis build badge
