+++ README.md
@@ -1,3 +1,5 @@
+[![Build Status](https://travis-ci.org/haptork/csaransh.svg?branch=master)](https://travis-ci.org/haptork/csaransh)
+
 Csaransh
 ==================================
 
