+++ csaransh-pp/.codecov.yml
@@ -1,0 +1,5 @@
+ignore:
+  - "third_party"  # ignore folders and all its contents
+  - "src/main.cpp"       # wildcards accepted
+  - "src/printJson.cpp"       # wildcards accepted
+  - "include/printJson.hpp"       # wildcards accepted
