+++ csaransh-pp/codecov.yml
@@ -1,5 +1,5 @@
 ignore:
-  - "csaransh-pp/third_party/*"  # ignore folders and all its contents
-  - "csaransh-pp/src/main.cpp"       # wildcards accepted
-  - "csaransh-pp/src/printJson.cpp"       # wildcards accepted
-  - "csaransh-pp/include/printJson.hpp"       # wildcards accepted
+  - "csaransh-pp/third_party/*" # ignoring catch and disjoint-set
+  - "csaransh-pp/src/main.cpp"  # ignoring for only boilerplate 
+  - "csaransh-pp/src/printJson.cpp" # ignoring for only printing boilerplate 
+  - "csaransh-pp/include/printJson.hpp" # ignoring for only printing boilerplate 
