### 2019-05-18 #1433 @haptork
@katyhuff I have added unit tests with edge cases to the C++ algorithms in the directory `csaransh-pp/test` with more than 600 assertions. I have also integrated travis CI and codecov. The code coverage badge shows around 90% coverage if we exclude boilerplate in main and printing of json data. The exercise turned out really helpful and I think it can also help in understanding of the behaviour of algorithm implementations. Thank you for making it part of the submission.
