### 2019-05-28 c314339
Refactored columns selection for table & stats
