+++ csaransh-server/views/js/cascades/MainTable.js
@@ -75,65 +75,20 @@ const RangeFilter = props => {
 
 const defaultRangeFilterFn = (filter, row) => row[filter.id] >= filter.value.min && row[filter.id] < filter.value.max;
 
-const minMaxSubc = (ar) => {
-  let min = 0, max = 0;
-  if (ar.length > 0) {
-    const val = Object.keys(ar[0].dclust_coords).length;
-    min = parseInt(val) == 0 ? 0 : parseInt(val) - 1;
-    max = min;
-  }
-  for (const x of ar) {
-    const val = Math.max(0, parseInt(Object.keys(x.dclust_coords).length) - 1);
-    min = Math.min(min, val);
-    max = Math.max(max, val);
-  }
-  return {min:Math.floor(min), max:Math.ceil(max + 0.01)};
-}; 
+const accessorDefault = name => x => x[name];
+const accessorOned = x => parseFloat(x["eigen_var"][0]) * 100;
+const accessorTwod = x => parseFloat(x["eigen_var"][0]) + parseFloat(x["eigen_var"][1]) * 100;
+const accessorSubc = x => (Object.keys(x.dclust_coords).length) <= 1 ? 0 : (Object.keys(x.dclust_coords).length);
 
-const minMaxOned = (ar) => {
+const minMaxPropsMaker = fields => (ar, name) => {
   let min = 0, max = 0;
   if (ar.length > 0) {
-    min = parseFloat(ar[0]["eigen_var"][0]) * 100;
-    max = min;
-  }
-  for (const x of ar) {
-    const val = parseFloat(x["eigen_var"][0]) * 100;
-    min = Math.min(min, val);
-    max = Math.max(max, val);
-  }
-  return {min:Math.floor(min), max:Math.ceil(max + 0.01)};
-}; 
-
-const minMaxTwod = (ar) => {
-  let min = 0, max = 0;
-  if (ar.length > 0) {
-    min = parseInt((parseFloat(ar[0]["eigen_var"][0]) + parseFloat(ar[0]["eigen_var"][1])) * 100);
-    max = min;
-  }
-  for (const x of ar) {
-    const val = parseInt((parseFloat(x["eigen_var"][0]) + parseFloat(x["eigen_var"][1])) * 100);
-    min = Math.min(min, val);
-    max = Math.max(max, val);
-  }
-  return {min:Math.floor(min), max:Math.ceil(max + 0.01)};
-};
-
-const minMaxProps = (ar, name) => {
-  if (name === "oned") {
-    return minMaxOned(ar);
-  } else if (name === "twod") {
-    return minMaxTwod(ar);
-  } else if (name === "subc") {
-    return minMaxSubc(ar);
-  }
-  let min = 0, max = 0;
-  if (ar.length > 0) {
-    const val = parseFloat(ar[0][name]);
+    const val = fields[name].accessor(ar[0]);
     min = val;
     max = val;
   }
   for (const x of ar) {
-    const val = parseFloat(x[name]);
+    const val = fields[name].accessor(x);
     min = Math.min(min, val);
     max = Math.max(max, val);
   }
@@ -154,8 +109,38 @@ const uniqueAr = (ar, key) => {
 }
 
 export class MainTable extends React.Component {
-  constructor(props) {
+ constructor(props) {
     super(props);
+    const roundOff = x => +parseFloat(x).toFixed(2);
+    const defaultParse = parseInt;
+    const defaultType = "output";
+    const both = (f, g) => x => f(g(x));
+    this.fields = {
+      "n_defects":{"name":"Defects Count"},
+      "max_cluster_size":{"name":"Max cluster size"}, 
+      "in_cluster":{"name":"% defects in cluster"}, 
[…]
