+++ presentation/pages/grapho_cascade/cascades_evolve.js
@@ -1,0 +1,2549 @@
+var cascades_evolve = 
+[[[78.3596, 73.8392, 54.9403, 1, 1, 1],
+  [76.0213, 74.9706, 55.7768, 1, 1, 1],
+  [73.7583, 67.6356, 57.8866, 1, 1, 1],
+  [75.874, 68.4722, 59.2663, 1, 1, 1],
+  [70.1705, 77.2901, 58.4533, 1, 1, 1],
+  [72.6816, 78.3752, 58.3848, 1, 1, 1],
+  [76.0901, 77.3722, 56.3633, 1, 1, 1],
+  [76.672, 64.3718, 64.3897, 1, 1, 1],
+  [70.2403, 71.2821, 61.4996, 1, 1, 1],
+  [73.3977, 70.8669, 61.4205, 1, 1, 1],
+  [71.1545, 70.623, 63.6546, 1, 1, 1],
+  [73.2467, 74.7706, 64.9399, 1, 1, 1],
+  [78.8089, 70.8301, 61.6213, 1, 1, 1],
+  [78.5606, 73.6216, 61.6355, 1, 1, 1],
+  [76.6866, 71.9511, 63.2741, 1, 1, 1],
+  [76.3911, 70.7666, 61.0707, 1, 1, 1],
+  [78.4701, 72.9872, 82.9413, 1, 1, 1],
+  [78.423, 63.2234, 77.4046, 1, 1, 1],
+  [72.3195, 56.1628, 65.9945, 1, 1, 1],
+  [73.4887, 58.201, 67.2704, 1, 1, 1],
+  [77.1118, 61.0858, 67.3862, 1, 1, 1],
+  [75.2467, 58.2155, 68.6918, 1, 1, 1],
+  [78.6635, 62.6025, 69.0235, 1, 1, 1],
+  [71.7012, 69.0154, 70.0442, 1, 1, 1],
+  [76.2775, 71.4984, 67.8724, 1, 1, 1],
+  [65.3281, 72.7158, 66.2692, 1, 1, 1],
+  [67.5648, 70.4234, 67.9467, 1, 1, 1],
+  [68.3703, 72.873, 66.7958, 1, 1, 1],
+  [68.7391, 72.428, 69.2399, 1, 1, 1],
+  [73.5923, 69.9662, 67.812, 1, 1, 1],
+  [71.0103, 74.139, 66.9435, 1, 1, 1],
+  [71.0494, 72.3177, 65.4781, 1, 1, 1],
+  [74.1304, 75.5735, 68.0547, 1, 1, 1],
+  [72.7804, 73.6598, 69.0792, 1, 1, 1],
+  [75.1085, 74.0135, 70.1655, 1, 1, 1],
+  [69.1487, 76.7959, 69.4901, 1, 1, 1],
+  [71.6475, 76.6283, 66.7567, 1, 1, 1],
+  [74.3302, 78.506, 68.0837, 1, 1, 1],
+  [72.525, 76.0598, 70.5859, 1, 1, 1],
+  [73.6981, 60.9676, 74.8661, 1, 1, 1],
+  [69.0193, 65.9111, 73.5267, 1, 1, 1],
+  [73.3594, 64.5558, 70.4099, 1, 1, 1],
+  [78.3943, 60.3188, 71.3328, 1, 1, 1],
+  [75.4525, 62.8386, 73.767, 1, 1, 1],
+  [77.7098, 64.2386, 75.2332, 1, 1, 1],
+  [75.4984, 64.5867, 71.9833, 1, 1, 1],
+  [70.4424, 67.3349, 71.5989, 1, 1, 1],
+  [74.2739, 68.1717, 70.2713, 1, 1, 1],
[…]
