+++ README.md
@@ -9,7 +9,8 @@ Csaransh
 The sofware suite to post-process, explore and visualize Molecular Dynamics (MD) simulations of collision cascades.
 It is an elaborate software solution for studing MD results of radiation damage simulations, starting from identifying defects from xyz file to finding correlations, visualizing subcascades to pattern matching clusters and the list goes on. There are many novel methods. The algorithms employed are fast and there is a refreshing interactive webapp to explore the results. You can check the web page for the project ([link](https://haptork.github.io/csaransh/)) and read the [manual](https://github.com/haptork/csaransh#getting-started) to go through the different sections of the analysis. The web page shows the results for the MD [database of IAEA Challenge on Materials for Fusion 2018](https://www-amdis.iaea.org/CDB/challenge/).
 
-For detailed discussion on new methods and results on classification of clusters please refer to this [paper](https://arxiv.org/abs/1811.10923).
+For detailed discussion on new methods and results on classification of clusters please refer to this [paper](https://arxiv.org/abs/1811.10923). Check the [interactive slides](https://haptork.github.io/csaransh/presentation/index.html) presented at [MoD-PMI 2019](http://dpc.nifs.ac.jp/dkato/MoD-PMI2019/).
+
 
 ## Affiliations and Acknowledgements
 
