### 2019-07-05 50a86ab
Added numba with version in conda env
