### 2019-07-05 9a374ce
Added correlations, minor fixes on small datasets
