+++ csaransh-server/views/js/cascades/CascadeVisualizer3D.js
@@ -342,7 +342,7 @@ export class CascadeVisualizer3D extends React.Component {
           this.all_frames[this.last_frame_name]["boxsize"] = boxsize;
           var energy = this_frame_data["energy"];
           var frequency = this_frame_data["clusterSizes"];
-          var classes = this_frame_data["clusterClasses"];
+          var classes = this_frame_data.hasOwnProperty("clusterClasses") ? this_frame_data["clusterClasses"] : {};
           var clusterids = new Set();
           var points_dict = {};
           coords.forEach(c => {
