### 2019-07-05 f0dc8b6
Updated with directions on testing, csaransh-app with gh-pages
