### 2019-07-08 ddcadcc
Update README.md
Minor fixes on help with web-app with gh-pages branch
