### 2019-07-08 #1461 @haptork
@katyhuff There seems to be no progress on this?

P.S. [here](https://haptork.github.io/csaransh/presentation/index.html) is a talk recently given on one of the analysis that is included in csaransh. The html presentation also uses the react components from csaransh web interface.
