### 2019-07-14 17dd8af
Added optimization flag if not running coverage
