### 2019-08-07 #1461 @jmborr
I am running the instructions in [README.md](https://github.com/haptork/csaransh/blob/master/README.md) in a Ubuntu 16.04 box. This I was able to do:
- [x] build the C++ post-processor and run the after-build tests
- [x] run `csaransh_pp` and `pp.py` on `data/lammps` to produce `cascades-data.js`
- [x] run the interface as a web application with node server. Here _instructions should state_ to `npm install csaransh-server` prior to `npm start`.

The duplication of this widget within section "View and Compare Cascade Info" is by design?
