### 2019-08-09 #1461 @haptork
> I am running the instructions in [README.md](https://github.com/haptork/csaransh/blob/master/README.md) in a Ubuntu 16.04 box. This I was able to do:
>
> * [x]  build the C++ post-processor and run the after-build tests
> * [x]  run `csaransh_pp` and `pp.py` on `data/lammps` to produce `cascades-data.js`
> * [x]  run the interface as a web application with node server. Here _instructions should state_ to `npm install csaransh-server` prior to `npm start`.
>
> The duplication of this widget within section "View and Compare Cascade Info" is by design?
> <img alt="" width="400" src="https://user-images.githubusercontent.com/1136006/62650322-9ae26400-b924-11e9-9cff-0748b872d918.png">

Thank you for taking this up. :)

Yes, the duplication is meant for comparison. The widget has a lock on top left. With lock open, it will show the stats for currently viewing cascade while once locked the stats get fixed on the current one selected. Thus, one cascade can be fixed on one of the widget and its stats can be compared with others as one views different cascades on the other widget with open lock. By default, the left widget is locked on some arbitrary cascade and right one would change as one selects the cascade from the main table. May be in future we can introduce some better interface to compare figures of two cascades but for now this does the job.

Please let me know of any other clarification if needed :)
