### 2019-08-21 #1461 @arose
> Installation: Does installation proceed as outlined in the documentation?

The installation went smoothly. I agree with @jmborr that `npm install csaransh-server` should be added to the instructions.

> References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

There are a few dois missing.
