### 2019-08-21 #1461 @arose
The IAEA challenge shows a need for visualization and analysis tools for collisional cascade molecular dynamics simulation data. Are the results of the challenge available somewhere? I could not find it on their website?
