### 2019-08-21 #1461 @katyhuff
Thanks @arose for your review!

Thank you @jmborr for your review efforts so far. Perhaps the remainder of your review could be revisited at this point (checkboxes at the top of the page).

@haptork, can you please respond (with changes to the code, install instructions, etc.) to the requests by the reviewers so far (i.e. the comment regarding adding the npm install command to the instructions, the comment from @rose regarding the dois and the results of the iaea challenge).
