### 2019-08-22 6114548
added note to run npm install before npm start.
