### 2019-08-23 #1461 @haptork
Thank you @arose for reviewing.

> > Installation: Does installation proceed as outlined in the documentation?
>
> The installation went smoothly. I agree with @jmborr that `npm install csaransh-server` should be added to the instructions.

Added to the instructions with latest commit (6114548d4da347d386cefa20da28d8db67bea651)

> The IAEA challenge shows a need for visualization and analysis tools for collisional cascade molecular dynamics simulation data. Are the results of the challenge available somewhere? I could not find it on their website?

The results were e-mailed to the participants. The results and the software tools of the participants are planned to be put up on the https://cascadesdb.org/ website which is a work in progress as of now.

The database that the contest uses and the one which is shown by default in csaransh does have an indirect reference. The database page of the contest is https://www-amdis.iaea.org/CDB/challenge/, for each data point it has a webpage (https://www-amdis.iaea.org/CDB/challenge/data/001.html), which shows the DOI of the publication for the datapoint. The reference publication for the data is http://doi.org/10.1038/s41467-018-03415-5 which we have added as a reference in paper.bib.

I don’t think there is any other reference for it now. We can share the mail notifying the results. Also, MoD-PMI presentation participation mentioned in paper, now has a link in the MoD-PMI web page ( http://dpc.nifs.ac.jp/dkato/MoD-PMI2019/ ) .

Let me know if anything needs to be done from our side to get the references checked.
