### 2019-09-09 b6d057a
Added four missing DOIs
