+++ paper.bib
@@ -4,7 +4,9 @@
         EDITOR = {Rudy J M Konings},
         BOOKTITLE = {Comprehensive Nuclear Materials},
         PUBLISHER = {Elsevier},
-        YEAR   = {2012}}
+        YEAR   = {2012},
+        DOI = {https://doi.org/10.1016/b978-0-08-056033-5.00027-6}
+        }
 @ARTICLE{BjorkasKai,
         AUTHOR = {C Bj\"orkas and K Nordlund and M J Caturla},
         TITLE  = {Influence of the picosecond defect distribution on damage accumulation in irradiated $\alpha$-Fe},
@@ -79,7 +81,8 @@ year="2013",
 publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="160--172",
-isbn="978-3-642-37456-2"
+isbn="978-3-642-37456-2",
+doi="https://doi.org/10.1007/978-3-642-37456-2_14"
 }
 @article{UMAP,
   title={Umap: Uniform manifold approximation and projection for dimension reduction},
@@ -104,7 +107,8 @@ journal = {Journal of Machine Learning Research}
   number={4},
   pages={46003},
   url={http://stacks.iop.org/0295-5075/103/i=4/a=46003},
-  year={2013}
+  year={2013},
+  doi={https://doi.org/10.1209/0295-5075/103/46003}
 }
 @article{Nordlund,
   title={Improving atomic displacement and replacement calculations with physically realistic damage models},
@@ -114,7 +118,8 @@ journal = {Journal of Machine Learning Research}
   number={1},
   pages={1084},
   year={2018},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={https://doi.org/10.1038/s41467-018-03415-5}
 }
 @online{reactjs,
   author = {Facebook},
