### 2019-09-09 c1300cf
Added two missing references
