+++ paper.bib
@@ -152,3 +152,16 @@ journal = {Journal of Machine Learning Research}
   url = {https://www-amdis.iaea.org/Workshops/ICTP2018/Presentations/ICTP2018TalkHill.pdf},
   urldate = {2018-10-09}
 }  
+@article{ubclasses,
+  title={Pattern Matching and Classification of Clusters in Collision Cascades},
+  author={Bhardwaj, Utkarsh and Sand, Andrea E and Warrier, Manoj},
+  journal={arXiv preprint arXiv:1811.10923},
+  year={2018}
+}
+@online{modpmi,
+  author = {Bhardwaj, Utkarsh and Sand, Andrea E and Warrier, Manoj},
+  publisher={http://dpc.nifs.ac.jp/dkato/MoD-PMI2019/},
+  title = {Classification of Clusters in MD Simulations of Collision Cascades},
+  url = {https://haptork.github.io/csaransh/presentation/index.html},
+  urldate = {2019-07-07}
+}
