+++ paper.md
@@ -37,7 +37,7 @@ The micro-structural properties of the materials change due to irradiation. The 
 - Different 3D-visualizations for a cascade including heat-maps, sub-cascades view, clusters-view etc. 
 - Interactive tool for pattern matching of cluster structures across the database.
 - Charts for comparison of properties of cascades such as cluster size distribution, spatial and angular distribution of defects from primary knock-on position etc. 
-- Statistical analysis over elements and energies for all the cascades for the properties like number of defects, dimensionality of cascades, number of sub-cascades etc.
+- Statistical analysis over elements and energies and correlations for all the cascades for the properties like number of defects, dimensionality of cascades, number of sub-cascades etc.
 - Interactive tool for exploring the classes identified for the clusters found.
 
 With the combination of efficient algorithms, unsupervised machine learning and modern interactive GUI with 3D visualizations the application helps in exploring different aspects of collision cascades qualitatively as well as quantitatively. The study can reveal many interesting correlations and patterns specific to different materials, energy ranges etc. We use efficient algorithm to identify defects from big MD simulation files. We use statistics and various unsupervised machine learning algorithms like HDBSCAN [@HDBSCAN], UMAP [@UMAP], PCA etc. to find various features such as identification of sub-cascades, characterization and classification of cluster structures, identifying dimensionality of cascades and clusters.
@@ -51,31 +51,31 @@ The ``CSaransh`` application was an entry in the IAEA challenge on materials for
 # Screenshots
 
 ![drawing](docs/table.png)
-*Main table with reactive filters, sorting, selection etc. The cascades of interest can be chosen here using the first (view) action button placed in first column of each row. Filters and sorting can be applied using the sliders for PKA energy, directions, defects count, etc. The statistics are shown for only the filtered rows. To exclude a single row from the statistics, the second action button for that row can be toggled.*
+*Main table with reactive filters, sorting, selection etc. The cascades of interest can be chosen here using the first (view) action button placed in first column of each row. Filters and sorting can be applied using the sliders for energy, directions, defects count, etc. The statistics are shown for only the filtered rows. To exclude a single row from the statistics, the second action button for that row can be toggled.*
 
 ![drawing](docs/clusterCmp.png)
-*Each cascade can be viewed as different sub-cascades. The sub-cascades are found using the DBSCAN algorithm on the vacancies created in the cascade.*
+*Section to chose and compare similar clusters. The cluster of interest can be chosen from the cascade view or the list on the left. The right side shows top five clusters from the complete database of cascades that match the shape of the cluster on the left. There are different matching criteria to choose from. Each of the matches can be separately viewed by clicking on the choices displayed on the bottom of the right section along with the (dis-)similarity metric from the cluster on left and dimensionality.*
 
 ![drawing](docs/subcascade1.png)
-*Each cascade can be viewed as different sub-cascades. The sub-cascades are found using the DBSCAN algorithm on the vacancies created in the cascade.*
+*A high energy cascade can be composed of different sub-cascades or frangments. The sub-cascades are found using the DBSCAN algorithm on the vacancies created in the cascade.*
 
 ![drawing](docs/subcascade2.png)
 *Scatter view and heat-map view for the same cascade. Heat map can be viewed for the vacancies, interstitials and both. It represents the density of defects. The scatter view and specifically heat map for vacancies can give a qualitative idea of the sub-cascading and range of the cascade.*
 
 ![drawing](docs/clusterSize.png)
-*Cluster size distribution of the selected cascades.*
+*Cluster size distribution of the cascades selected from the main table.*
 
 ![drawing](docs/ndefects.png)
[…]
