### 2019-09-09 2f6b6a8
Removed extra line above image caption
