### 2019-09-09 #1461 @haptork
@katyhuff
Added missing DOIs (commit b6d057ab6d21a7bed34e90737d10d7b98e170402)
Added missing refs (commit c1300cf3547740809255e08250c6a0d9e3862436)
Updated screenshot captions, removed extra line between image and caption(commit 2f6b6a8626fe128cf55a27002bc24b44d4713d1e)
