### 2019-09-10 dbe2f2c
Moved image captions to placeholders
