+++ paper.md
@@ -50,31 +50,22 @@ The ``CSaransh`` application was an entry in the IAEA challenge on materials for
 
 # Screenshots
 
-![drawing](docs/table.png)
-*Main table with reactive filters, sorting, selection etc. The cascades of interest can be chosen here using the first (view) action button placed in first column of each row. Filters and sorting can be applied using the sliders for energy, directions, defects count, etc. The statistics are shown for only the filtered rows. To exclude a single row from the statistics, the second action button for that row can be toggled.*
+![Main table with reactive filters, sorting, selection etc. The cascades of interest can be chosen here using the first (view) action button placed in first column of each row. Filters and sorting can be applied using the sliders for energy, directions, defects count, etc. The statistics are shown for only the filtered rows. To exclude a single row from the statistics, the second action button for that row can be toggled.](docs/table.png)
 
-![drawing](docs/clusterCmp.png)
-*Section to chose and compare similar clusters. The cluster of interest can be chosen from the cascade view or the list on the left. The right side shows top five clusters from the complete database of cascades that match the shape of the cluster on the left. There are different matching criteria to choose from. Each of the matches can be separately viewed by clicking on the choices displayed on the bottom of the right section along with the (dis-)similarity metric from the cluster on left and dimensionality.*
+![Section to chose and compare similar clusters. The cluster of interest can be chosen from the cascade view or the list on the left. The right side shows top five clusters from the complete database of cascades that match the shape of the cluster on the left. There are different matching criteria to choose from. Each of the matches can be separately viewed by clicking on the choices displayed on the bottom of the right section along with the (dis-)similarity metric from the cluster on left and dimensionality.](docs/clusterCmp.png)
 
-![drawing](docs/subcascade1.png)
-*A high energy cascade can be composed of different sub-cascades or frangments. The sub-cascades are found using the DBSCAN algorithm on the vacancies created in the cascade.*
+![A high energy cascade can be composed of different sub-cascades or frangments. The sub-cascades are found using the DBSCAN algorithm on the vacancies created in the cascade.](docs/subcascade1.png)
 
-![drawing](docs/subcascade2.png)
-*Scatter view and heat-map view for the same cascade. Heat map can be viewed for the vacancies, interstitials and both. It represents the density of defects. The scatter view and specifically heat map for vacancies can give a qualitative idea of the sub-cascading and range of the cascade.*
+![Scatter view and heat-map view for the same cascade. Heat map can be viewed for the vacancies, interstitials and both. It represents the density of defects. The scatter view and specifically heat map for vacancies can give a qualitative idea of the sub-cascading and range of the cascade.](docs/subcascade2.png)
 
-![drawing](docs/clusterSize.png)
-*Cluster size distribution of the cascades selected from the main table.*
+![Cluster size distribution of the cascades selected from the main table.](docs/clusterSize.png)
 
-![drawing](docs/ndefects.png)
-*Number of defects over different energies in Fe, On mouse-over the boxes show mean, median, max etc. Box plot for different properties can be plotted using the drop down menu. The statistics are aggregated for the rows filtered from the main table.*
+![Number of defects over different energies in Fe, On mouse-over the boxes show mean, median, max etc. Box plot for different properties can be plotted using the drop down menu. The statistics are aggregated for the rows filtered from the main table.](docs/ndefects.png)
 
-![drawing](docs/splom.png)
-*The scatter plot of each parameter against every other is plotted that helps in checking correlations and trends between the parameters such as number of defects, number of sub-cascades, in cascade clustering fraction etc. The color coding is for number of defects, which helps to see if a pair of parameters correlate to the number of defects.*
[…]
