+++ paper.bib
@@ -125,32 +125,37 @@ journal = {Journal of Machine Learning Research}
   author = {Facebook},
   title = {Reactjs},
   url = {https://reactjs.org},
-  urldate = {2018-10-09}
+  urldate = {2018-10-09},
+  year={2013}
 }
 @online{chartjs,
   author = {Chartjs},
   title = {Ceactjs},
   url = { http://www.chartjs.org/},
-  urldate = {2018-10-09}
+  urldate = {2018-10-09},
+  year={2014}
 }
 @online{plotlyjs,
   author = {Plotly},
   title = {Plotly.js},
   url = {https://plot.ly/javascript/},
-  urldate = {2018-10-09}
+  urldate = {2018-10-09},
+  year={2013}
 }
 @online{IAEA-challenge,
   author = {IAEA},
   title = {IAEA Challenge on Materials for Fusion - 2018},
   url = {https://challenge.iaea.org/challenges/2018-NA-Mat-Fusion},
-  urldate = {2018-10-09}
+  urldate = {2018-10-09},
+  year={2018}
 }
 @online{CascadesDB,
   author = {Christian Hill},
   publisher={IAEA},
   title = {Managing Large Atomic and Molecular Data Sets: HITRAN, ExoMol and CascadesDB},
   url = {https://www-amdis.iaea.org/Workshops/ICTP2018/Presentations/ICTP2018TalkHill.pdf},
-  urldate = {2018-10-09}
+  urldate = {2018-10-09},
+  year={2018}
 }  
 @article{ubclasses,
   title={Pattern Matching and Classification of Clusters in Collision Cascades},
@@ -163,5 +168,6 @@ journal = {Journal of Machine Learning Research}
   publisher={http://dpc.nifs.ac.jp/dkato/MoD-PMI2019/},
   title = {Classification of Clusters in MD Simulations of Collision Cascades},
   url = {https://haptork.github.io/csaransh/presentation/index.html},
-  urldate = {2019-07-07}
+  urldate = {2019-07-07},
+  year={2019}
 }
