### 2019-09-10 #1461 @haptork
Updated a few formatting issues in the latest commit (putting captions as image placeholders).
