### 2019-09-10 92b555f
Minor fixes for formatting
