### 2019-09-11 #1461 @katyhuff
@haptork Thanks for engaging actively in the review process. I've made some suggestions on the pdf itself (from before your most recent commit). Can you please consider the suggestions marked in this pdf, and let me know as soon as you feel these suggested changes have been handled?

[csaransh.pdf](https://github.com/openjournals/joss-reviews/files/3598409/csaransh.pdf)
