### 2019-09-11 f79d24f
Updated as per comments of JOSS editor
