### 2019-09-11 #1461 @katyhuff
Thank you @arose and @jmborr for your reviews -- we couldn't do this without you.
Thank you @haptork for a strong submission and for engaging actively in the review process! I have looked over the paper, double-checked all the DOI links, and have conducted a high-level review of the code itself. Everything looks ship-shape to me. At this point, please double-check the paper yourself, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service.

**Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission.**
Please also let me know if the release version needs to be updated in this JOSS submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission. Until then, now is your moment for final touchups!
