### 2019-09-12 6ecb2ba
Corrected a few typos
