### 2019-09-15 #1461 @katyhuff
@openjournals/joss-eics : I believe this is ready for your review
