### 2019-09-16 #1461 @arfon
@openjournals/joss-eics : I believe this is ready for your review

/ cc @kthyng
