### 2019-09-16 #1461 @kthyng
Hi @haptork! A few items to wrap up your submission:

- [ ] The tagged version in the github repo is v0.3.4, the release version is v0.3.3, and the version just mentioned in this conversation (and listed at the top) is v0.3.2. Can you confirm the version and make these consistent?

Edits for paper:
 - [ ] pg. 2 paragraph 2: should "materials for fusion, 2018" be capitalized?
 - [ ] pg. 2 paragraph 2: "in arXiv preprint paper (Bhardwaj, Sand, & Warrier, 2018)" should be "in *the*..." and is it important to specify that it is a preprint? Was it never published in a peer-reviewed journal?
 - [ ] pg.2 paragraph 3: "CascadesDB database(Hill, 2018)" needs a space
 - [ ] "etc" is used repeatedly throughout the paper — please remove these except when strictly necessary.
 - [ ] Figure 2: "Section to chose and compare similar clusters" should be "Section to choose and compare similar clusters"
 - [ ] Figure 2: "The right side shows top five clusters" should be "The right side shows the top five clusters"
 - [ ] Figure 3: "A high energy cascade can be composed of different sub-cascades or frangments." has a typo "fragments"
