### 2019-09-17 d9d3cb9
Updated paper.md as per final reviewer.
