+++ paper.md
@@ -30,29 +30,29 @@ bibliography: paper.bib
 
 # Summary
 
-The micro-structural properties of materials change due to irradiation. The defects formed during the displacement cascades caused by irradiation are the primary source of radiation damage [@Stoller; @BjorkasKai]. The number of primary defects produced, defect cluster size distribution, and defect cluster structures after a collision cascade can be studied using Molecular Dynamics simulations. These results determine the long term evolution of the micro-structural changes in the material [@Stoller; @GOLUBOV; @SINGH; @OSETSKY; @BECQUART]. The properties studied using Molecular Dynamics (MD) simulations can be used in higher scale radiation damage models like Monte Carlo methods and rate theories etc. [@OSETSKY; @BECQUART].
+The micro-structural properties of materials change due to irradiation. The defects formed during the displacement cascades caused by irradiation are the primary source of radiation damage [@Stoller; @BjorkasKai]. The number of primary defects produced, defect cluster size distribution, and defect cluster structures after a collision cascade can be studied using Molecular Dynamics simulations. These results determine the long term evolution of the micro-structural changes in the material [@Stoller; @GOLUBOV; @SINGH; @OSETSKY; @BECQUART]. The properties studied using Molecular Dynamics (MD) simulations can be used in higher scale radiation damage models like Monte Carlo methods and rate theories. [@OSETSKY; @BECQUART].
 
 ``CSaransh`` is a software suite to explore the Molecular Dynamics simulations of collision cascades. It includes post-processors to identify defects, characterize and classify cluster-structures, find number of sub-cascades etc. and a single page web-application (SPA) as GUI that provides interactive visualizations and charts such as:
 
-- Different 3D-visualizations for a cascade including heat-maps, sub-cascades view, clusters-view etc. 
+- Different 3D-visualizations for a cascade including heat-maps, sub-cascades view, clusters-view. 
 - Interactive tool for pattern matching of cluster structures across the database.
-- Charts for comparison of properties of cascades such as cluster size distribution, spatial and angular distribution of defects from primary knock-on position etc. 
+- Charts for comparison of properties of cascades such as cluster size distribution, spatial and angular distribution of defects from primary knock-on position. 
 - Statistical analysis over elements and energies and correlations for all the cascades for the properties like number of defects, dimensionality of cascades, number of sub-cascades etc.
 - Interactive tool for exploring the classes identified for the clusters found.
 
-With the combination of efficient algorithms, unsupervised machine learning and modern interactive GUI with 3D visualizations the application helps in exploring different aspects of collision cascades qualitatively as well as quantitatively. Using the application many interesting correlations and patterns specific to different materials, energy ranges etc. can be explored. We developed an efficient algorithm to identify defects from big MD simulation files. We use statistics and various unsupervised machine learning algorithms like HDBSCAN [@HDBSCAN], UMAP [@UMAP], PCA etc. to find various features such as identification of sub-cascades, characterization and classification of cluster structures based on features we have developed, identifying dimensionality of cascades and clusters.
+With the combination of efficient algorithms, unsupervised machine learning and modern interactive GUI with 3D visualizations the application helps in exploring different aspects of collision cascades qualitatively as well as quantitatively. Using the application many interesting correlations and patterns specific to different materials or energy ranges can be explored. We developed an efficient algorithm to identify defects from big MD simulation files. We use statistics and various unsupervised machine learning algorithms like HDBSCAN [@HDBSCAN], UMAP [@UMAP], PCA to find various features such as identification of sub-cascades, characterization and classification of cluster structures based on features we have developed, identifying dimensionality of cascades and clusters.
 
[…]
