### 2019-09-17 #1461 @haptork
> Hi @haptork! A few items to wrap up your submission:
>
> * [ ]  The tagged version in the github repo is v0.3.4, the release version is v0.3.3, and the version just mentioned in this conversation (and listed at the top) is v0.3.2. Can you confirm the version and make these consistent?

The zenodo archive version for the DOI I mentioned is 0.3.2 (in its metadata) which corresponds to: https://github.com/haptork/csaransh/releases/tag/v0.3.2. Please ignore the later two versions in the GitHub repo. Sorry for the confusion.

> Edits for paper:
>
> * [ ]  pg. 2 paragraph 2: should "materials for fusion, 2018" be capitalized?

Thank you. Corrected.

> * [ ]  pg. 2 paragraph 2: "in arXiv preprint paper (Bhardwaj, Sand, & Warrier, 2018)" should be "in _the_..." and is it important to specify that it is a preprint? Was it never published in a peer-reviewed journal?

Thank you. Added _the_ and removed preprint from the sentence. The paper is in communication in a peer reviewed journal.

> * [ ]  pg.2 paragraph 3: "CascadesDB database(Hill, 2018)" needs a space

Thanks. Corrected.

> * [ ]  "etc" is used repeatedly throughout the paper — please remove these except when strictly necessary.

Corrected. Thanks.

> * [ ]  Figure 2: "Section to chose and compare similar clusters" should be "Section to choose and compare similar clusters"

Thanks. It had already been corrected before archiving.

> * [ ]  Figure 2: "The right side shows top five clusters" should be "The right side shows the top five clusters"

Thanks. Corrected

> * [ ]  Figure 3: "A high energy cascade can be composed of different sub-cascades or frangments." has a typo "fragments"

Thanks. It had already been corrected before archiving.

commit for the corrections: d9d3cb9f6e84b737ba22dfd5064d8a545cbb5cc7
zenodo archive after the corrections (if required): 10.5281/zenodo.3434635

Please let me know if anything is to be done.
