### 2019-09-19 #1461 @kthyng
Great! Thanks @haptork for addressing those items.
